Improve nbfs integration.
Sending SIGHUP to the daemon now makes it reload the server spec file.
This allows to change the set of exposed features without restarting
the daemon.
However, note that an image has to be un-mounted & remounted before
a change in the features can be taken into account.

The server now always updates nbfs exports file (when triggered by
a change in image mounts) if the nbfs feature is declared in the spec
file. Before it was only if nbfsd PID was found. Then a SIGHUP is
sent to nbfsd daemon if possible.

+++ server/walt/server/threads/main/network/nbfs.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 import os, signal
+from walt.server.spec import server_has_feature
 
 NBFSD_PID_PATH = Path('/var/lib/nbfs/nbfsd.pid')
 NBFS_EXPORTS_PATH = Path('/etc/nbfs/exports')
@@ -29,10 +30,11 @@ def get_nbfsd_pid():
         return None
 
 def update_exports(root_paths, subnet):
-    nbfsd_pid = get_nbfsd_pid()
-    if nbfsd_pid is None:
-        # nbfsd probably not installed or running
-        return
-    generate_exports_file(root_paths, subnet)
-    # notify nbfsd it should re-read its exports file
-    os.kill(nbfsd_pid, signal.SIGHUP)
+    if server_has_feature('nbfs'):
+        generate_exports_file(root_paths, subnet)
+        nbfsd_pid = get_nbfsd_pid()
+        if nbfsd_pid is None:
+            # nbfsd probably not installed or running
+            return
+        # notify nbfsd it should re-read its exports file
+        os.kill(nbfsd_pid, signal.SIGHUP)
